refactor(Services): AddBed, RoomBed to own service
