+++ hotel-booking-app/Services/IBedService.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using HotelBookingApp.Pages;
 
 namespace HotelBookingApp.Services
 {
@@ -11,5 +12,6 @@ namespace HotelBookingApp.Services
     {
         Task<IEnumerable<SelectListItem>> FindAll();
         Task<Bed> FindBedById(int bedId);
+        Task<RoomBed> AddBed(BedViewModel model);
     }
 }
